+++ elastic_cloud_agent/tools/__init__.py
@@ -1,11 +1,9 @@
 """Tools for the Elastic Cloud Agent."""
 
-from elastic_cloud_agent.tools.openapi_tool import create_openapi_toolkit
 from elastic_cloud_agent.tools.search import create_search_tool
 from elastic_cloud_agent.tools.smart_openapi_toolkit import create_smart_openapi_toolkit
 
 __all__ = [
-    "create_openapi_toolkit",
     "create_smart_openapi_toolkit",
     "create_search_tool",
 ]
