+++ elastic_cloud_agent/tools/smart_openapi_toolkit.py
@@ -6,9 +6,11 @@ tool smarter - it classifies user intent and filters the OpenAPI specification
 to only show relevant endpoints, reducing token usage and improving accuracy.
 """
 
+import json
 from pathlib import Path
 from typing import Any, Dict, List, Optional
 
+import yaml
 from langchain_community.agent_toolkits.json.base import create_json_agent
 from langchain_community.agent_toolkits.json.toolkit import JsonToolkit
 from langchain_community.agent_toolkits.openapi.prompt import DESCRIPTION
@@ -19,12 +21,50 @@ from langchain_core.language_models import BaseLanguageModel
 from langchain_core.tools import BaseTool, Tool
 from langchain.agents.agent import AgentExecutor
 
-from elastic_cloud_agent.tools.openapi_tool import create_requests_wrapper, load_api_spec
+from elastic_cloud_agent.utils.config import Config, get_api_spec_path
 from elastic_cloud_agent.utils.openapi_utils import (
     classify_intent,
     extract_tags_from_spec,
     filter_spec_by_tags,
 )
+
+
+def load_api_spec(spec_path: Optional[Path] = None) -> Dict[str, Any]:
+    """
+    Load and parse the OpenAPI specification.
+
+    Args:
+        spec_path: Path to the OpenAPI specification file. If None, use the default path.
+
+    Returns:
+        Dict[str, Any]: The parsed OpenAPI specification
+    """
+    if spec_path is None:
+        spec_path = get_api_spec_path()
+
+    with open(spec_path, "r") as f:
+        if spec_path.suffix in (".yaml", ".yml"):
+            return yaml.safe_load(f)
+        elif spec_path.suffix == ".json":
+            return json.load(f)
+        else:
+            raise ValueError(f"Unsupported file format: {spec_path.suffix}")
+
+
+def create_requests_wrapper() -> TextRequestsWrapper:
+    """
+    Create a requests wrapper for the Elastic Cloud API.
+
+    Returns:
+        TextRequestsWrapper: A configured requests wrapper
+    """
+    # Configure headers with authentication
+    headers = {
+        "Content-Type": "application/json",
+        "Authorization": f"ApiKey {Config.ELASTIC_CLOUD_API_KEY}",
+    }
+
+    return TextRequestsWrapper(headers=headers)
 
 
 class SmartJsonExplorerTool:
